Delete and Update Task Done by User

+++ src/tasks/tasks.service.ts
@@ -5,7 +5,7 @@ import { TaskRepository } from './tasks.repository'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Task } from './tasks.entity'
 import { TaskStatus } from './task-status-enum'
-import { User } from '../auth/user.entity';
+import { User } from '../auth/user.entity'
 
 @Injectable()
 export class TasksService {
@@ -13,12 +13,14 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository
   ) {}
 
-  async getTasks(filterDto: GetTaskFilterDto, user: User) : Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto, user);
+  async getTasks (filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
   async getTaskById (id: string, user: User): Promise<Task> {
-    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } })
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id }
+    })
     if (!found) {
       throw new NotFoundException(`Task with ID '${id}' not found`)
     }
@@ -26,24 +28,25 @@ export class TasksService {
     return found
   }
 
-  async createTask (
-    createTaskDto: CreateTaskDto,
-    user: User,
-    ): Promise<Task> {
+  async createTask (createTaskDto: CreateTaskDto, user: User): Promise<Task> {
     return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async deleteTask (id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
+  async deleteTask (id: string, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID '${id}' not found`)
     }
   }
 
-  // async updateTaskStatus(id: number, status: TaskStatus) : Promise<Task> {
-  //   const task = await this.getTaskById(id);
-  //   task.status = status;
-  //   await task.save();
-  //   return task;
-  // }
+  async updateTaskStatus (
+    id: string,
+    status: TaskStatus,
+    user: User
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
+    task.status = status
+    await task.save()
+    return task
+  }
 }
